refactor: deepen FintualPerformance and collapse snapshot folding and persistence
